rewrite: rewrote component declaration and separated from component expression
